dashboard api keys

+++ src/commands/deleteMessages.ts
@@ -1,43 +1,33 @@
-// @ts-nocheck
-//TODO investigate this file
-
-const {
-  SlashCommandBuilder,
-  CommandInteraction,
-  Client,
-} = require("discord.js");
-const { getDiscordServer } = require("../functions/prisma/discord");
+import { SlashCommandBuilder, CommandInteraction, Client } from "discord.js";
+import { getDiscordServer } from "../functions/prisma/discord";
+import { prisma } from "../index";
 
 module.exports = {
   developer: true,
   data: new SlashCommandBuilder()
     .setName("delete-messages")
     .setDescription("Batch deletes messages (developer only)"),
-  /**
-   *
-   * @param {CommandInteraction} interaction
-   * @param {Client} client
-   */
-  async execute(interaction, client) {
+  async execute(interaction: any, client: any) {
     await interaction.reply("Deleting...");
     setTimeout(async () => await interaction.deleteReply(), 5000);
 
     const guildID = BigInt(interaction.guildId);
-    const prisma = require("../index");
     const server = await getDiscordServer(prisma, guildID);
 
-    for (const faction of server.factions) {
-      const channel = await interaction.guild.channels.cache.get(
-        faction.discordChannelId.toString()
-      );
-
-      // delete all possible messages
-      await channel.bulkDelete(100, true).then(() => {
-        console.log(
-          "Messages deleted of ",
+    if (server) {
+      for (const faction of server.factions) {
+        const channel = await interaction.guild.channels.cache.get(
           faction.discordChannelId.toString()
         );
-      });
+
+        // delete all possible messages
+        await channel.bulkDelete(100, true).then(() => {
+          console.log(
+            "Messages deleted of ",
+            faction.discordChannelId.toString()
+          );
+        });
+      }
     }
   },
 };
